Moves Entrance schema
